Removed unused imports and moved NettyServer.java to the root of the server package
