+++ client/src/main/java/com/porejemplo/nube/client/service/DownloadService.java
@@ -1,7 +1,6 @@
 package com.porejemplo.nube.client.service;
 
 import java.io.IOException;
-import java.nio.file.Path;
 
 public interface DownloadService {
     boolean download(String fileName) throws IOException;
