+++ server/src/main/java/com/porejemplo/nube/server/auth/repository/UserDAO.java
@@ -14,8 +14,6 @@ public class UserDAO {
         try {
             Class.forName("org.sqlite.JDBC");
             connection = DriverManager.getConnection("jdbc:sqlite:server\\mainDb.db");
-            // Tests in AuthServiceTest.java will run successfully if you uncomment below line and comment the above line.
-//            connection = DriverManager.getConnection("jdbc:sqlite:mainDb.db");
             stmt = connection.createStatement();
         } catch (ClassNotFoundException e) {
             e.printStackTrace();
