+++ server/src/main/java/com/porejemplo/nube/server/netty/AuthStateOfAuthHandler.java
@@ -1,12 +1,6 @@
 package com.porejemplo.nube.server.netty;
 
-import com.porejemplo.nube.common.Command;
-import com.porejemplo.nube.server.auth.service.AuthService;
-import io.netty.buffer.ByteBuf;
-import io.netty.buffer.ByteBufAllocator;
 import io.netty.channel.ChannelHandlerContext;
-
-import java.nio.charset.StandardCharsets;
 
 public class AuthStateOfAuthHandler implements State {
