+++ server/src/main/java/com/porejemplo/nube/server/netty/DownloadCommandReceivedStateOfMainHandler.java
@@ -1,7 +1,6 @@
 package com.porejemplo.nube.server.netty;
 
 import com.porejemplo.nube.common.Command;
-import io.netty.buffer.ByteBuf;
 import io.netty.buffer.ByteBufAllocator;
 import io.netty.channel.*;
