+++ server/src/main/java/com/porejemplo/nube/server/netty/Phase.java
@@ -9,7 +9,6 @@ public enum Phase {
     VERIFY_FILE_PRESENCE,
     FILE_DESPATCH,
     FILES_LIST,
-    NAME_AND_NEW_NAME_LENGTH,
     NAME_AND_NEW_NAME,
     RENAME_FILE,
     NEW_NAME_LENGTH,
