+++ server/src/main/java/com/porejemplo/nube/server/netty/State.java
@@ -1,6 +1,5 @@
 package com.porejemplo.nube.server.netty;
 
-import io.netty.buffer.ByteBuf;
 import io.netty.channel.ChannelHandlerContext;
 
 import java.io.IOException;
